Fixed warnings
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4899 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/geometry/Contour2D.cpp
@@ -536,8 +536,8 @@ namespace
 
             const Vector3D<float> a1((float)p(0), (float)p(1), 0);
             const Vector3D<float> a2((float)q(0), (float)q(1), 0);
-            const Vector3D<float> b1((float)p(0), (float)p(1), height);
-            const Vector3D<float> b2((float)q(0), (float)q(1), height);
+            const Vector3D<float> b1((float)p(0), (float)p(1), (float)height);
+            const Vector3D<float> b2((float)q(0), (float)q(1), (float)height);
 
             const Vector3D<float> normal = normalize(cross(a2 - a1, b1 - a1));
